apr_jwe_decrypt_content_aesgcm() null terminate string
Also fixes unit test failure.

The test test_jwt_decrypt_gcm() in test.c would sometimes fail. The
failure was caused by extra garbage appearing in the decrypted string
after the expected string. This is due to the failure of
apr_jwe_decrypt_content_aesgcm() to null terminate the string as is
similarity done in apr_jwe_decrypt_content_aescbc().

This patch copies the logic from apr_jwe_decrypt_content_aescbc()
regarding the decryption output length (p_len) and the final output
length (f_len) which are then used to compute the end of the data in
the output buffer, the next octet after the decrypted data in the
output buffer is then assigned a null terminator.

Signed-off-by: John Dennis <jdennis@redhat.com>