handle "X-Requested-With: XMLHttpRequest" as non-browser; closes #228
Previously clients sending a "X-Requested-With:" header would be treated
as non-browser clients and would receive a 401 on unauthenticated
requests. As it turns out this would block Android WebViews from
accessing protected content. Now only those clients sending the specific
case-insensitive "XMLHttpRequest" value in the "X-Requested-With:"
header will be treated as such. Thanks @mguillem.

Signed-off-by: Hans Zandbelt <hans.zandbelt@zmartzone.eu>