Do not delete the session cookie in the event of a cache backend failure
The cache backend will always return TRUE in the event of a
authoritative cache hit or miss. FALSE is returned only in cases where
the cache backend is unable answer the query or mod_auth_openidc is
unable to grab a global lock. By not deleting the session cookie in the
cases where the backend has failed, we allow clients to gracefully
resume once the backend is restored.